Hey there, world
We're a small team of award-winning local devs and tech mentors bringing our community into the next era by educating at all levels of computer literacy. Looking to join in on amazing projects or simply to hone your skills? Say hello here and enroll in a session or three!

This (so far) one-page hello to the world - just like each of our past events - was designed and built from the ground up as a collaboration between our students and tutors! Come, learn with us, too!